requirements.txt
SmartApi-python
logzero
pyotp
pandas>=2.0
numpy
python-dotenv
pytz
requests
websocket-client
ib_insync==0.9.86
